Separate view components and its rendering into separate directories
