+++ app/helpers/view_component/course_card_component.rb
@@ -3,7 +3,7 @@
 module ViewComponent
   module CourseCardComponent
     def course_card_component(course:, enrollment: nil)
-      render partial: 'view_components/app/course_carousal/course_card_component', locals: { course:, enrollment: }
+      render partial: 'view_components/course_carousal/course_card_component', locals: { course:, enrollment: }
     end
   end
 end
