+++ app/helpers/view_component/course_carousal_component.rb
@@ -3,12 +3,12 @@
 module ViewComponent
   module CourseCarousalComponent
     def course_carousal_component(courses:, title:, count:, load_path:)
-      render partial: 'view_components/app/course_carousal/course_carousal_component',
+      render partial: 'view_components/course_carousal/course_carousal_component',
              locals: { courses:, title:, count:, load_path: }
     end
 
     def course_carousal_body_component(courses:)
-      render partial: 'view_components/app/course_carousal/course_carousal_body_component', locals: { courses: }
+      render partial: 'view_components/course_carousal/course_carousal_body_component', locals: { courses: }
     end
   end
 end
