+++ app/views/ui/components/app.html.erb
