+++ app/views/ui/icons/index.html.erb
