+++ app/views/ui/notification_bars/index.html.erb
@@ -1,3 +1,7 @@
+<div class="my-6">
+  <%= render "shared/components/back_button", back_link: ui_components_path %>
+</div>
+
 <div class="flex flex-col gap-4">
   <div class="heading heading-medium pb-6">Banners</div>
   <%= notification_bar(
