+++ app/views/ui/tables/index.html.erb
