+++ app/views/view_components/course_carousal/_course_card_component.html.erb
