+++ app/views/view_components/course_carousal/_course_carousal_body_component.html.erb
