+++ app/views/view_components/course_carousal/_course_carousal_component.html.erb
