+++ app/views/view_components/course_select/_course_select_component.html.erb
