+++ app/views/view_components/course_select/_list_component.html.erb
