+++ app/views/view_components/course_select/_list_item_component.html.erb
