+++ app/views/view_components/course_select/_search_component.html.erb
