+++ app/views/view_components/course_select/_sidebar_component.html.erb
