+++ app/views/view_components/notification_bar/_notification_bar.html.erb
